+++ dart/lib/sentry.dart
@@ -34,7 +34,7 @@ export 'src/sentry_envelope.dart';
 export 'src/sentry_envelope_item.dart';
 export 'src/sentry_options.dart';
 // ignore: invalid_export_of_internal_element
-export 'src/sentry_span_operations.dart';
+export 'src/constants.dart';
 // ignore: invalid_export_of_internal_element
 export 'src/sentry_trace_origins.dart';
 export 'src/span_data_convention.dart';